% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_clients.R
\name{available_clients}
\alias{available_clients}
\title{Available clients}
\usage{
available_clients()
}
\description{
<ToDo> Note to self: provided names are case independent (here and at send_message)
}
\seealso{
\link{custom_client}
}
