% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_gmail.R
\name{client_gmail}
\alias{client_gmail}
\alias{send_message.client_gmail}
\title{Gmail client}
\usage{
client_gmail(email, key, secret)

\method{send_message}{client_gmail}(client, message, destination,
  verbose = FALSE, decode_response = TRUE,
  subject = "notifieR notification", ...)
}
\arguments{
\item{email}{Email of reciver of the message.}

\item{key}{\code{Key} created in Google API application.}

\item{secret}{\code{Secret Key} created in Google API application}

\item{client}{TODO}

\item{message}{TODO}

\item{destination}{TODO}

\item{verbose}{TODO}

\item{decode_response}{TODO}

\item{subject}{Subject of email message.}
}
\description{
Client extending a \link{client_notifieR} for Gmail service.
In addition to any fields in the \link{client_notifieR} this one contains
\code{email},\code{key},\code{secret} which is needed to send a message via
Gmail Send API.

\link{send_message} method for \code{client_gmail}. TODO
}
