% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_notifieR.R
\name{client_notifieR}
\alias{client_notifieR}
\alias{is.client_notifieR}
\alias{send_message.client_notifieR}
\alias{print.client_notifieR}
\title{notifieR client}
\usage{
client_notifieR(service)

is.client_notifieR(x)

\method{send_message}{client_notifieR}(client, message, destination,
  verbose = FALSE, decode_response = TRUE, ...)

\method{print}{client_notifieR}(x, ...)
}
\arguments{
\item{service}{a name of a client's serivce.}

\item{client}{TODO}

\item{message}{TODO}

\item{destination}{TODO}

\item{verbose}{TODO}

\item{decode_response}{TODO}
}
\description{
Function creates an object of \code{client_notifieR} class with
given \code{service} field, typically a name of selected service. This object
is an underlying structure from which every other client is created.

\code{is.client_notifieR} checks if a provided object is of 
the \code{client_notifieR} class and whether it has all the fields
a \code{client_notifieR} should have.

\link{send_message} method for \code{client_notifieR} only
serves as a placeholder. For details on how to create a custom client
see \link{custom_client}.
}
\seealso{
\link{available_clients}, \link{custom_client}
}
