% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_gmail.R, R/client_notifieR.R,
%   R/client_slack.R, R/client_telegram.R
\name{is.client_gmail}
\alias{is.client_gmail}
\alias{is.client_notifieR}
\alias{is.client_slack}
\alias{is.client_telegram}
\title{notifieR clients' verification}
\usage{
is.client_gmail(x)

is.client_notifieR(x)

is.client_slack(x)

is.client_telegram(x)
}
\arguments{
\item{x}{object to be tested.}
}
\description{
\code{is.client_notifieR} checks if a provided object is of 
the \code{client_notifieR} class and whether it has all the fields
a \code{client_notifieR} should have.

All other functions check if the provided object extends
\code{client_notifieR} and whether it has all the fields a given client
should have.
}
