% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_gmail.R, R/client_notifieR.R,
%   R/client_slack.R, R/client_telegram.R
\name{is.client_gmail}
\alias{is.client_gmail}
\alias{is.client_sendeR}
\alias{is.client_slack}
\alias{is.client_telegram}
\title{sendeR clients' verification}
\usage{
is.client_gmail(x)

is.client_sendeR(x)

is.client_slack(x)

is.client_telegram(x)
}
\arguments{
\item{x}{an object to be tested.}
}
\description{
\code{is.client_sendeR} tests if a provided object is of
the \code{client_sendeR} class and whether it has all the fields
a \code{client_sendeR} should have.

All other functions check if the provided object extends
\code{client_sendeR} (passes \code{is.client_sendeR} test) and whether it
has all the fields a given client should have.
}
