% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_gmail.R, R/client_notifieR.R,
%   R/client_slack.R, R/client_telegram.R, R/send_message.R
\name{send_message.client_gmail}
\alias{send_message.client_gmail}
\alias{send_message.client_notifieR}
\alias{send_message.client_slack}
\alias{send_message.client_telegram}
\alias{send_message}
\title{Methods for sending messages.}
\usage{
\method{send_message}{client_gmail}(client, message, destination,
  verbose = FALSE, subject = "notifieR notification", ...)

\method{send_message}{client_notifieR}(client, message, destination,
  verbose = FALSE, ...)

\method{send_message}{client_slack}(client, message, destination = NULL,
  verbose = FALSE, ...)

\method{send_message}{client_telegram}(client, message, destination,
  verbose = FALSE, ...)

send_message(client, message, destination, verbose = FALSE, ...)
}
\arguments{
\item{client}{TODO}

\item{message}{TODO}

\item{destination}{TODO}

\item{verbose}{TODO}

\item{subject}{a subject of an email message.}

\item{...}{TODO}
}
\description{
The \code{send_message} method for the \code{client_notifieR}
only serves as a placeholder.

TODO
}
