% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_notifieR.R
\name{client_notifieR}
\alias{client_notifieR}
\alias{print.client_notifieR}
\title{notifieR client}
\usage{
client_notifieR(service)

\method{print}{client_notifieR}(x, ...)
}
\arguments{
\item{service}{typically a name of service whose client extends the object.}

\item{x}{an object to print.}

\item{...}{argument not used. Required only for overloading purposes.}
}
\description{
Function \code{client_notifieR} is a constructor for an object
of a \code{client_notifieR} class. This object is an underlying structure
on which every other client is based.
}
\seealso{
\link{is.client_notifieR}, \link{send_message},
}
