% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_sendeR.R
\name{client_sendeR}
\alias{client_sendeR}
\alias{print.client_sendeR}
\title{sendeR client}
\usage{
client_sendeR(service)

\method{print}{client_sendeR}(x, ...)
}
\arguments{
\item{service}{typically a name of service whose client extends the object.}

\item{x}{an object to print.}

\item{...}{argument not used. Required only for overloading purposes.}
}
\description{
Function \code{client_sendeR} is a constructor for an object
of a \code{client_sendeR} class. This object is an underlying structure
on which every other client is based.
}
\seealso{
\link{is.client_sendeR}, \link{send_message},
}
