% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_slack.R
\name{client_slack}
\alias{client_slack}
\title{Slack client}
\usage{
client_slack(slack_webhook)
}
\arguments{
\item{slack_webhook}{a webhook obtained from the Slack API settings.}
}
\description{
Client extending the \link{client_notifieR} for the Slack
service. In addition to any fields in the \link{client_notifieR} this one
contains \code{slack_webhook} which is needed to send a message via the Slack
Webhook API. For additional information on how to create a webhook see details.
}
\details{
To create your own webhook head to
\url{https://api.slack.com/messaging/webhooks}.
\strong{Note}: Webhooks are permamently connected to one channel.
}
\seealso{
\link{is.client_slack}, \link{send_message},
}
