% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_telegram.R
\name{client_telegram}
\alias{client_telegram}
\alias{is.client_telegram}
\alias{send_message.client_telegram}
\title{Telegram client}
\usage{
client_telegram(telegram_token)

is.client_telegram(x)

\method{send_message}{client_telegram}(client, message, destination,
  verbose, ...)
}
\arguments{
\item{telegram_token}{<ToDo>}
}
\description{
Client extending a \link{notifieR_client} for Telegram service.
In addition to any fields in the \link{notifieR_client} this one contains
\code{telegram_token} which is needed to send a message via Telegram API.

\code{is.client_telegram} checks if a provided object is of 
the \code{client_telegram} class, whether it has all the fields
a \code{client_telegram} should have and if it is a \link{notifieR_client}.

\link{send_message} method for \code{client_telegram}. <ToDo>
}
