% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_telegram.R
\name{client_telegram}
\alias{client_telegram}
\title{Telegram client}
\usage{
client_telegram(telegram_token)
}
\arguments{
\item{telegram_token}{a bot token given by the Botfather.}
}
\description{
Client extending a \link{client_sendeR} for the Telegram
service. In addition to any fields in the \link{client_sendeR} this one
contains \code{telegram_token} which is needed to send a message via
the Telegram Bot API. For additional information on how to get required
\code{telegram_token} see details.
}
\details{
To create your own telegram bot refer to
\url{https://core.telegram.org/bots#6-botfather}.
}
\seealso{
\link{is.client_telegram}, \link{send_message}
}
