% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_telegram.R
\name{client_telegram}
\alias{client_telegram}
\title{Telegram client}
\usage{
client_telegram(telegram_token, ...)
}
\arguments{
\item{telegram_token}{a bot token given by the BotFather.}

\item{...}{named arguments with additional fields which will be passed to
\code{\link{set_fields}} during client creation.}
}
\description{
Client extending a \code{\link{client_sendeR}} for the Telegram
service. In addition to any fields in the \code{\link{client_sendeR}} this one
contains \code{telegram_token} which is needed to send a message via
the Telegram Bot API. For additional information on how to get required
\code{telegram_token} see details.
}
\details{
To create your own telegram bot refer to
\url{https://core.telegram.org/bots#6-botfather}.
}
\examples{
client <- client_telegram("my_token")

# Variant with default parameters set
client2 <- client_telegram("my_token", message = "Default message template")

}
\seealso{
\code{\link{is.client_telegram}}, \code{\link{send_message}}
}
