% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_gmail.R, R/client_notifieR.R,
%   R/client_slack.R, R/client_telegram.R, R/send_message.R
\name{send_message.client_gmail}
\alias{send_message.client_gmail}
\alias{send_message.client_notifieR}
\alias{send_message.client_slack}
\alias{send_message.client_telegram}
\alias{send_message}
\title{Methods for sending messages.}
\usage{
\method{send_message}{client_gmail}(client, message, destination,
  verbose = FALSE, subject = "notifieR notification", ...)

\method{send_message}{client_notifieR}(client, message, destination,
  verbose = FALSE, ...)

\method{send_message}{client_slack}(client, message, destination = NULL,
  verbose = FALSE, ...)

\method{send_message}{client_telegram}(client, message, destination,
  verbose = FALSE, ...)

send_message(client, message, destination, verbose = FALSE, ...)
}
\arguments{
\item{client}{one of the clients created by the notifieR package.}

\item{message}{a text body of the message to send. Should be a character
vector of length one.}

\item{destination}{a destination of the message. Should be a character
vector of length one. For more details see \strong{Methods (by class)}
section below.}

\item{verbose}{if is set to \code{TRUE} then the function returns whole
response instead of only a status_code.}

\item{subject}{a subject of an email message.}

\item{...}{\code{send_message} passes additional (non-standard) arguments to
respective method. Client specific methods have this argument only for
method overloading purposes.}
}
\description{
The \code{send_message} method for the \code{client_notifieR}
only serves as a placeholder.
}
\section{Methods (by class)}{
\itemize{
\item \code{client_gmail}: \code{destination} is a recipient's email address.

\item \code{client_slack}: \code{description} is a ... TODO(TK)

\item \code{client_telegram}: \code{destination} is a recipient's chat id.
}}

