% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_fields.R
\name{set_fields}
\alias{set_fields}
\title{Set fields inside a sendeR client}
\usage{
set_fields(client, ...)
}
\arguments{
\item{client}{The client which will have default values set.}

\item{...}{Named arguments to set inside the client (non-named arguments will
be ignored).}
}
\value{
Updated client.
}
\description{
This function is intended to set default parameters such as 
\code{message} or \code{destination} inside a client as additional fields.
After setting default values the clients will be able to send messages
without passing additional arguments to the \code{\link{send_message}} function.
}
\details{
Standard fields utilized in the package include:
\itemize{
    \item \code{message}
    \item \code{destination} (apart from Slack service client)
    \item \code{verbose}
}
any other set defaults won't have any effect.
}
\examples{
client <- client_telegram("my_token")
client <- set_fields(client, message = "Hello world!", destination = 12345)

}
